Add creation date filters to SearchPagesRequest and update converter.

+++ Apps.Hubspot/Utils/Converters/DateTimeToUnixEpoch.cs
@@ -12,7 +12,7 @@ public class DateTimeToUnixEpoch : JsonConverter
     public override void WriteJson(JsonWriter writer, object value, JsonSerializer serializer)
     {
         var dateTime = (DateTime)value;
-        writer.WriteValue(((int)dateTime.Subtract(DateTime.UnixEpoch).TotalSeconds).ToString());
+        writer.WriteValue(((long)dateTime.Subtract(DateTime.UnixEpoch).TotalSeconds).ToString());
     }
 
     public override object ReadJson(JsonReader reader, Type objectType, object existingValue, JsonSerializer serializer)
